Add support for closures
The idea is to use upvalues to capture variables from the surrounding
scopes. The upvalues serve as an indirection to the actual variables,
allowing the closure to access variables that are not in its own scope.
When the captured variable goes out of scope, we close-over the
variable. This is what Lua does. See [crafting interpreters]
(https://craftinginterpreters.com/closures.html) for more details.

I've also replaced `Vec` with `ArrayVec` for the VM stack to allow the
upvalues to maintain a pointer to the local in the stack frame. Using
`Vec` would have made the pointers invalid when the stack grew or shrank,
so we need to use a fixed-size array.