Add a mark and sweep garbage collector
- The collector is a most basic version of mark and sweep collection
and is triggered when the allocated bytes exceed a certain threshold.
The default threshold is 1MB which changes as the GC collects garbage
and the allocated bytes change
- Fix a bug in the `new` method of `Closure` where the upvalues vector
should not be initialized, otherwise the GC would try to dereference
null pointers
- Add a `sizeof` function to the VM that returns the size of types that
the VM and GC use. This is used to estimate the size of the heap
currently in use by the program